pyproject.toml
[tool.poetry]
name = "ocr-with-tesseract"
version = "0.1.0"
description = ""
authors = ["VeerSinghCurl <85334331+VeerSinghCurl@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9.2"
opencv-python = "^4.5.2"
pytesseract = "^0.3.7"
pdf2image = "^1.15.1"
cmake = "^3.20.2"
python-poppler = "^0.2.2"
jsons = "^1.4.2"
python-Levenshtein = "^0.12.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
